Fixed to address compiler warnings
git-svn-id: https://svn.code.sf.net/p/openpfgw/code@661 368e28a5-49e5-40c5-be6c-d84989c25d7b

+++ pform/pfio/pfwin32guioutput.cpp
@@ -18,7 +18,7 @@
 
 char g_cpTrayMsg[200] = "\0";
 
-PFWin32GUIOutput::PFWin32GUIOutput(long long hWnd) : PFOutput(), m_hWnd(hWnd)
+PFWin32GUIOutput::PFWin32GUIOutput(uint64_t hWnd) : PFOutput(), m_hWnd(hWnd)
 {
 }
